+++ lib/authlogic.rb
@@ -15,7 +15,6 @@ path = File.dirname(__FILE__) + "/authlogic/"
 [
   "i18n",
   "random",
-  "regex",
   "config",
 
   "controller_adapters/abstract_adapter",